+++ app/grouping/[id]/members/components/MemberItem.client.tsx
@@ -1,3 +1,4 @@
+'use client';
 import { Spacing } from '@/components/common/Spacing';
 import clsx from 'clsx';
 import Image from 'next/image';
